Attempt to fix merge conflict
